Add support for forcing 6 bpc on DP pipes.
For DP compliance we need to be able to control the output color type
for the pipe associated with the DP port.  To do this we rely on the
intel_dp_test_force_bpc sysfs file and the associated value stored in
struct intel_dp.  If the sysfs file has a non-zero value and we're on a
display port connector, then we use the value from sysfs to calculate the
bpp for the pipe.  For cases where we are not on DP or there has not been
an overridden value then we behave as normal.

Signed-off-by: Jim Bride <jim.bride@linux.intel.com>